+++ src/client/rsg-components/Markdown/List/List.spec.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { render } from 'enzyme';
 
 import List from './index';
